Fixing invalid host problem
If an invalid host was provided, it would try to throw an error that
was not yet defined (ShelfError).

Along with this I noticed we were not running coverage on index which
is why a test hadn't caught this.

Also changed where error stuff was stored. It COULD go on the shelfLib
but then you wouldn't be able to use error.INVALID_HOST_PREFIX since you
would not yet have access to a shelfLib.