Required for uploading very large files
from a AWS's network. The exact reason is still unknown to me. It
appears that what happens is the file gets done uploading to the
shelf-api machine but then while it is waiting for a response
(and shelf-api is uploading to S3) it goes on for longer than
60 seconds without receiving or sending any packets. If we are
supposed to see keep alive packets, we are not.