Required for uploading very large files
from a very fast network. The exact reason why is still unknown to
me but it appears that the connection stays open while the server
continues to acknowledge packets and at some point the client says
"ENOUGH", sends a [FIN, ACK] and a few [RST]s to the server.

This results in a socket timeout.